+++ packages/ethers5/exports/index.ts
@@ -2,7 +2,7 @@ import type { Web3ModalOptions } from '../src/client.js'
 import { Web3Modal } from '../src/client.js'
 import { ConstantsUtil } from '@web3modal/scaffold-utils'
 
-export type { Web3ModalOptions } from '../src/client.js'
+export type { Web3Modal, Web3ModalOptions } from '../src/client.js'
 export { defaultConfig } from '../src/utils/defaultConfig.js'
 
 export function createWeb3Modal(options: Web3ModalOptions) {