+++ packages/scaffold/src/partials/w3m-account-wallet-features-widget/styles.ts
@@ -21,4 +21,18 @@ export default css`
   wui-tabs {
     width: 100%;
   }
+
+  .contentContainer {
+    height: 280px;
+  }
+
+  .contentContainer > wui-icon-box {
+    width: 40px;
+    height: 40px;
+    border-radius: var(--wui-border-radius-xxs);
+  }
+
+  .contentContainer > .textContent {
+    width: 65%;
+  }
 `