+++ packages/scaffold/src/partials/w3m-input-address/styles.ts
@@ -1,0 +1,56 @@
+import { css } from 'lit'
+
+export default css`
+  :host {
+    width: 100%;
+    height: 100px;
+    border-radius: var(--wui-border-radius-s);
+    border: 1px solid var(--wui-gray-glass-002);
+    background-color: var(--wui-gray-glass-002);
+    transition: all var(--wui-ease-out-power-1) var(--wui-duration-lg);
+    position: relative;
+  }
+
+  :host(:hover) {
+    background-color: var(--wui-gray-glass-005);
+  }
+
+  wui-flex {
+    width: 100%;
+    height: fit-content;
+  }
+
+  wui-button {
+    display: ruby;
+    color: var(--wui-color-fg-100);
+    margin: 0 var(--wui-spacing-xs);
+  }
+
+  .instruction {
+    position: absolute;
+    top: 50%;
+    transform: translateY(-50%);
+  }
+
+  textarea {
+    background: transparent;
+    width: 100%;
+    font-family: var(--w3m-font-family);
+    font-size: var(--wui-font-size-medium);
+    font-style: normal;
+    font-weight: var(--wui-font-weight-light);
+    line-height: 130%;
+    letter-spacing: var(--wui-letter-spacing-medium);
+    color: var(--wui-color-fg-100);
+    caret-color: var(--wui-color-accent-100);
+    box-sizing: border-box;
+    -webkit-appearance: none;
+    -moz-appearance: textfield;
+    padding: 0px;
+    border: none;
+    outline: none;
+    appearance: none;
+    resize: none;
+    overflow: hidden;
+  }
+`