+++ packages/wagmi/exports/index.ts
@@ -2,7 +2,7 @@ import type { Web3ModalOptions, CoreConfig } from '../src/client.js'
 import { Web3Modal } from '../src/client.js'
 import { ConstantsUtil } from '@web3modal/scaffold-utils'
 
-export type { Web3ModalOptions } from '../src/client.js'
+export type { Web3Modal, Web3ModalOptions } from '../src/client.js'
 export { defaultWagmiConfig } from '../src/utils/defaultWagmiCoreConfig.js'
 export { emailConnector } from '../src/connectors/EmailConnector.js'